renaming the hooks script file
