+++ src/route/sign-in.jsx
@@ -4,7 +4,7 @@ import wastelessLogo from "../assets/brand/logo.svg";
 import Button from "../component/base/Button";
 import SignInForm from "../component/SignInForm";
 import { fireAuth } from "../firebase";
-import { useFirebaseAuthData } from "../firebase/hook";
+import { useFirebaseAuthData } from "../firebase/hooks";
 import { useError } from "../utils/hooks";
 
 export default function Page() {
